♻ Adjust browser download script
Pipe directly to the file, cleanup on error, and download a revision that exists for all platforms

+++ packages/core/src/utils/install-browser.js
@@ -1,4 +1,3 @@
-import os from 'os';
 import path from 'path';
 import https from 'https';
 import {
@@ -10,30 +9,36 @@ import rimraf from 'rimraf';
 import log from '@percy/logger';
 import readableBytes from './bytes';
 
+// used to determine platform defaults
+const platform = (
+  process.platform === 'win32' && process.arch === 'x64'
+    ? 'win64' : process.platform
+);
+
 export default async function install({
   // default discovery browser is chromium
   browser = 'Chromium',
-  // default chromium version is 78.0.3904.87
-  revision = '693954',
+  // default chromium version is 78.0.3904.x
+  /* istanbul ignore next: hard to cover sans combined reports */
+  revision = platform === 'win64' ? '693951' : '693954',
   // default download directory is in @percy/core root
   directory = path.resolve(__dirname, '../../.local-chromium'),
   // default download url is dependent on platform and revision
   url = `https://storage.googleapis.com/chromium-browser-snapshots/${{
     linux: `Linux_x64/${revision}/chrome-linux.zip`,
     darwin: `Mac/${revision}/chrome-mac.zip`,
-    win32: os.arch() === 'x64'
-      ? `Win_x64/${revision}/chrome-win64.zip`
-      /* istanbul ignore next: hard to cover sans combined reports */
-      : `Win/${revision}/chrome-win32.zip`
-  }[os.platform()]}`,
+    win64: `Win_x64/${revision}/chrome-win.zip`,
+    win32: `Win/${revision}/chrome-win32.zip`
+  }[platform]}`,
   // default extraction method is to unzip
   extract = (i, o) => require('extract-zip')(i, { dir: o }),
   // default exectuable location within the extracted archive
   executable = {
     linux: 'chrome',
+    win64: 'chrome.exe',
     win32: 'chrome.exe',
     darwin: path.join('Chromium.app', 'Contents', 'MacOS', 'Chromium')
-  }[os.platform()]
+  }[platform]
 } = {}) {
   let outdir = path.join(directory, revision);
   let dlpath = path.join(outdir, url.split('/').pop());
@@ -46,35 +51,46 @@ export default async function install({
 
     log.info(`${browser} not found, downloading...`);
 
-    // ensure the out directory exists
-    await fs.mkdir(outdir, { recursive: true });
+    try {
+      // ensure the out directory exists
+      await fs.mkdir(outdir, { recursive: true });
 
-    // download the file at the given URL and log progress
-    await new Promise((resolve, reject) => {
-      https.get(url, response => {
-        let file = createWriteStream(dlpath);
-        let size = parseInt(response.headers['content-length'], 10);
-        let msg = `${revision} - ${readableBytes(size)} [:bar] :percent :etas`;
-        let progress = new (require('progress'))(log.formatter(msg), {
-          stream: process.stdout,
-          incomplete: ' ',
-          total: size,
-          width: 20
-        });
+      // download the file at the given URL and log progress
+      await new Promise((resolve, reject) => {
+        https.get(url, response => {
+          if (response.statusCode !== 200) {
+            response.resume();
+            reject(new Error(`Download failed: ${response.statusCode} - ${url}`));
+            return;
+          }
 
-        response.on('data', chunk => {
-          file.write(chunk);
-          progress.tick(chunk.length);
-        }).on('end', () => {
-          file.end();
-          resolve();
-        });
-      }).on('error', reject);
-    });
+          let size = parseInt(response.headers['content-length'], 10);
+          let msg = `rev ${revision} - ${readableBytes(size)} [:bar] :percent :etas`;
+          let progress = new (require('progress'))(log.formatter(msg), {
+            stream: process.stdout,
+            incomplete: ' ',
+            total: size,
+            width: 20
+          });
 
-    // extract the downloaded file and cleanup
-    await extract(dlpath, outdir);
-    await new Promise(resolve => rimraf(dlpath, resolve));
+          let file = createWriteStream(dlpath)
+            .on('finish', resolve)
+            .on('error', reject);
+
+          response.on('data', chunk => {
+            progress.tick(chunk.length);
+          }).pipe(file);
+        }).on('error', reject);
+      });
+
+      // extract the downloaded file
+      await extract(dlpath, outdir);
+    } finally {
+      // always cleanup
+      if (existsSync(dlpath)) {
+        await new Promise(resolve => rimraf(dlpath, resolve));
+      }
+    }
 
     // log success and restore previous loglevel
     log.info(`Successfully downloaded ${browser}`);
